+++ src/components/CardInfoForm/InputContainers/CardNumbersInputContainer.tsx
@@ -1,8 +1,8 @@
-import Input from '../common/Input';
-import { ErrorWrapper, ErrorText } from '../../styles/common';
-import InputSection from '../common/InputSection';
-import getObjectKeys from '../../utils/getObjectKeys';
-import useDisplayingErrorStatus from '../../hooks/useDisplayingErrorStatus';
+import Input from '../../common/Input';
+import { ErrorWrapper, ErrorText } from '../../../styles/common';
+import InputContainer from '../../common/InputContainer';
+import getObjectKeys from '../../../utils/getObjectKeys';
+import useDisplayingErrorStatus from '../../../hooks/useDisplayingErrorStatus';
 
 type CardNumberKey = 'first' | 'second' | 'third' | 'fourth';
 
@@ -30,7 +30,7 @@ export default function CardNumbersContainer({ value, generateOnChange, errorSta
 
   return (
     <div>
-      <InputSection
+      <InputContainer
         title="결제할 카드 번호를 입력해 주세요"
         subtitle="본인 명의의 카드만 결제 가능합니다."
         labelText="카드 번호"
@@ -54,7 +54,7 @@ export default function CardNumbersContainer({ value, generateOnChange, errorSta
             />
           );
         })}
-      </InputSection>
+      </InputContainer>
       <ErrorWrapper>
         <ErrorText>{errorMessage}</ErrorText>
       </ErrorWrapper>
