+++ src/components/CardInfoForm/InputContainers/CvcInputContainer.tsx
@@ -1,7 +1,7 @@
-import { ErrorText, ErrorWrapper } from '../../styles/common';
-import Input from '../common/Input';
-import InputSection from '../common/InputSection';
-import useDisplayingErrorStatus from '../../hooks/useDisplayingErrorStatus';
+import { ErrorText, ErrorWrapper } from '../../../styles/common';
+import Input from '../../common/Input';
+import InputContainer from '../../common/InputContainer';
+import useDisplayingErrorStatus from '../../../hooks/useDisplayingErrorStatus';
 
 export interface ICvcInputContainerProps {
   value: string;
@@ -29,7 +29,7 @@ const CvcInputContainer = ({ value, onChange, errorStatus, setIsCardFront }: ICv
 
   return (
     <div>
-      <InputSection title="CVC 번호를 입력해 주세요" labelFor="CVC" labelText="CVC">
+      <InputContainer title="CVC 번호를 입력해 주세요" labelFor="CVC" labelText="CVC">
         <Input
           isError={isError}
           placeholder="123"
@@ -40,7 +40,7 @@ const CvcInputContainer = ({ value, onChange, errorStatus, setIsCardFront }: ICv
           onBlur={onBlur}
           width="100%"
         />
-      </InputSection>
+      </InputContainer>
       <ErrorWrapper>
         <ErrorText>{errorMessage}</ErrorText>
       </ErrorWrapper>
