+++ src/components/CardInfoForm/InputContainers/PasswordInputContainer.tsx
@@ -1,7 +1,7 @@
-import InputSection from '../common/InputSection';
-import { ErrorText, ErrorWrapper } from '../../styles/common';
-import Input from '../common/Input';
-import useDisplayingErrorStatus from '../../hooks/useDisplayingErrorStatus';
+import InputContainer from '../../common/InputContainer';
+import { ErrorText, ErrorWrapper } from '../../../styles/common';
+import Input from '../../common/Input';
+import useDisplayingErrorStatus from '../../../hooks/useDisplayingErrorStatus';
 
 export interface IPasswordInputContainerProps {
   value: string;
@@ -19,7 +19,7 @@ export default function PasswordInputContainer({ value, onChange, errorStatus }:
 
   return (
     <div>
-      <InputSection
+      <InputContainer
         title="비밀번호를 입력해 주세요"
         subtitle="앞의 2자리를 입력해 주세요"
         labelFor="password"
@@ -35,7 +35,7 @@ export default function PasswordInputContainer({ value, onChange, errorStatus }:
           onBlur={bringErrorStatus}
           placeholder="12"
         />
-      </InputSection>
+      </InputContainer>
       <ErrorWrapper>
         <ErrorText>{errorMessage}</ErrorText>
       </ErrorWrapper>
